% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Systematic review which compares the accuracy of HbA1c vs
FPG in deabetes}
\format{A data frame with 38 rows and 9 columns. Each row represents study results, the columns are:

   \describe{
\item{Study}{Name of the first author.}
\item{TP_HbA1c}{Number of true positive cases for HbA1c.}
\item{FP_HbA1c}{Number of false positive cases for HbA1c.}
\item{FN_HbA1c}{Number of false negative cases for HbA1c.}
\item{TN_HbA1c}{Number of true negative cases for HbA1c.}
\item{TP_FPG}{Number of true positive cases for FPG.}
\item{FP_FPG}{Number of false positive cases for FPG.}
\item{FN_FPG}{Number of false negative cases for FPG.}
\item{TN_FPG}{Number of true negative cases for FPG.}
    }}
\source{
Hoyer, A., Kuss, O. Meta-analysis for the comparison of two
disgnostic test: a new approach based on copulas. Stat. Med. 2018; 37:739-748
}
\description{
This data frame contains results of diagnostic accuraccy of 38 studies which reported comparison of sensitivity and specificity between HbA1c vs FPG in a population based screening for type 2 diabetes.
}
\details{
This data frame contains results of diagnostic accuraccy of 38 studies
which reported comparison of sensitivity and specificity between
HbA1c vs FPG in a population-based screening for type 2 diabetes.
}
\keyword{datasets}
