% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{ep}
\alias{ep}
\title{Ectopic pregnancy vs. all other pregnancies data}
\format{A matrix with 21 rows. and 8 columns. Each row represents study
results, the columns are:
\describe{
\item{    tp   }{true positives}
\item{    n1   }{number of patients with disease}
\item{    fp   }{false positives}
\item{    n2   }{number of patients without disease}
\item{    d1   }{Prospective vs retrospective}
\item{    d2   }{Cohort vs case-control}
\item{    d3   }{Consecutive sampling patients series vs non-consecutive} }}
\source{
Table III Mol et al. 1998
}
\description{
Ectopic pregnancy vs. all other pregnancies data
Table III Mol et al. 1998
}
\references{
Verde P. E. (2010). Meta-analysis of diagnostic test data: A
bivariate Bayesian modeling approach. \emph{Statistics in Medicine}.
\bold{29}, 3088-3102.
}
\keyword{datasets}

