% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{glas}
\alias{glas}
\title{Tumor markers in the diagnosis of primary bladder cancer.}
\format{A matrix with 46 rows and 7 columns. Each row represents study
results, the columns are:

\describe{
\item{tp}{number of true positives.}
\item{n1}{number of patients with disease.}
\item{fp}{number of false positives.}
\item{n2}{number of patients without disease.}
\item{author}{first author of the study.}
\item{cutoff}{cutoff in U/ml.}
 \item{marker}{test method used in the study.}
 }}
\source{
The data were obtained from

Glas AS, Roos D, Deutekom M, Zwindermann AH, Bossuyt PM, Kurth KH. (2003)
Tumor markers in the diagnosis of primary bladder cancer. A systematic
review. \emph{Journal of Urology}; \bold{169}:1975-82.
}
\description{
Outcome of individual studies evaluating urine markers
}
\references{
Verde P. E. (2010). Meta-analysis of diagnostic test data: A
bivariate Bayesian modeling approach. \emph{Statistics in Medicine}.
\bold{29}, 3088-3102.
}
\keyword{datasets}
