% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{gould}
\alias{gould}
\title{Accuracy of Positron Emission Tomography for Diagnosis of Pulmonary
Nodules and Mass Lesions}
\format{A matrix with 31 rows and 6 columns. Each row represents study
results, the columns are:
\describe{
\item{    tp   }{true positives}
\item{    n1   }{number of patients with disease}
\item{    fp   }{false positives}
\item{    n2   }{number of patients without disease}
\item{    author   }{first author of the study}
\item{    year   }{publication
 date} }}
\source{
The data were obtainded from

Gould MK, Maclean CC, Kuschner WG, Rydzak CE, Owens Dk. (2001) Accuracy of
positron emission tomography for diagnosis of pulmonary nodules and mass
lesions: a meta-analysis. \emph{The Journal of the American Medical
Association};\bold{285}:914-24.
}
\description{
Data from a Meta-Analysis of Studies on
Quality of Studies of FDG-PET for Diagnosis of SPNs and Mass Lesions
}
\references{
Verde P. E. (2010). Meta-analysis of diagnostic test data: A
bivariate Bayesian modeling approach. \emph{Statistics in Medicine}.
\bold{29}, 3088-3102.
}
\keyword{datasets}

