% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotsesp.R
\name{plotsesp}
\alias{plotsesp}
\title{plotsesp()
plot the posterior densities for Se and Sp}
\usage{
plotsesp(m, binwidth.p = 0.03, CI.level = 0.95)
}
\arguments{
\item{m}{The matrix used for plotting the diagram.}

\item{binwidth.p}{Histograms binwidth, default is 0.03.}

\item{CI.level}{Level of the posterior interval default is 0.95.}
}
\description{
plotsesp()
plot the posterior densities for Se and Sp
}
\examples{
## execute analysis
\dontrun{
data(ep)
m1.ep <- metadiag(ep[,1:4])

plotsesp(m = m1.ep)
}
}
\seealso{
\code{\link{metadiag}}.
}
\keyword{file}

