% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.metadiag.R
\name{print.metadiag}
\alias{print.metadiag}
\title{Generic print function for metadiag object in bamdit}
\usage{
\method{print}{metadiag}(x, digits = 3, ...)
}
\arguments{
\item{x}{The object generated by the function metadiag.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for metadiag object in bamdit
}
