% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{rapt}
\alias{rapt}
\title{Systematic reviews of clinical decision tools for acute abdominal pain}
\format{A data frame with 13 rows and 13 columns. Each row represents study results, the columns are:
   \describe{
\item{Author}{Name of the first author and year of publication}
\item{tp.dr}{Number of true positive cases for unadded doctors.}
\item{fp.dr}{Number of false positive cases for unadded doctors.}
\item{fn.dr}{Number of false negative cases for unadded doctors.}
\item{tn.dr}{Number of true negative cases for unadded doctors.}
\item{tp.tools}{Number of true positive cases for doctors with decision tools.}
\item{fp.tools}{Number of false positive cases for doctors with decision tools.}
\item{fn.tools}{Number of false negative cases for doctors with decision tools.}
\item{tn.tools}{Number of true negative cases for doctors with decision tools.}
\item{tool}{Diagnostic tool.}
\item{n.dr}{Total number of cases for unadded doctors.}
\item{n.tools}{Total number of cases for doctors with decision tools.}
\item{design}{Study design.}
    }}
\source{
Health Technol Assess. 2006 Nov;10(47):1-167, iii-iv.
Systematic reviews of clinical decision tools for acute abdominal pain.
Liu JL1, Wyatt JC, Deeks JJ, Clamp S, Keen J, Verde P, Ohmann C,
Wellwood J, Dawes M, Altman DG.
}
\description{
This data frame corresponds to 13 clinical studies reporting the accuracy of doctors added with decision tools.
}
\details{
This data frame contains results of diagnostic accuraccy of 13 studies
which reported comparison of sensitivity and specificity between
doctors using diagnostic tools vs doctors without decision tools.
}
\references{
Health Technol Assess. 2006 Nov;10(47):1-167, iii-iv.
Systematic reviews of clinical decision tools for acute abdominal pain.
Liu JL1, Wyatt JC, Deeks JJ, Clamp S, Keen J, Verde P, Ohmann C,
Wellwood J, Dawes M, Altman DG.
}
\keyword{datasets}
