\name{read.RevMan}
\alias{read.RevMan}
\title{Reads the Review Managers standart export file}
\description{
This function reads the data for the analysis given the location of the standart export file of the Review Manager application 
}
\usage{
read.RevMan(file)
}
\arguments{
  \item{file}{The file location}
}
\seealso{
 % \code{\link{bn}}.
	\url{http://ims.cochrane.org/revman}
}
\examples{


\dontrun{ read.RevMan(file)}

}
\keyword{file}