% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{scheidler}
\alias{scheidler}
\title{Radiological evaluation of lymph node metastases in patients with cervical cancer: a
meta-analysis.}
\format{A matrix with 46 rows and 7 columns. Each row represents study
results, the columns are:
\describe{
\item{    tp   }{true positives}
\item{    n1   }{number of patients with disease}
\item{    fp   }{false positives}
\item{    n2   }{number of patients without disease}
\item{    author   }{first author of the study}
\item{    year   }{publication date}
\item{    test   )}{test method used in the study} }}
\source{
The data were obtainded from

Scheidler J, Hricak H, Yu KK, Subak L, Segal MR. (1997) Radiological
evaluation of lymph node metastases in patients with cervical cancer: a
meta-analysis. \emph{The Journal of the American Medical Association};
\bold{278}:1096-1101.
}
\description{
This data frame summarizes the tables 1-3 of Scheidler et al. 1997.
}
\references{
Verde P. E. (2010). Meta-analysis of diagnostic test data: A
bivariate Bayesian modeling approach. \emph{Statistics in Medicine}.
\bold{29}, 3088-3102.
}
\keyword{datasets}

