% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{skin}
\alias{skin}
\title{Accuracy of Computer-Aided Diagnosis of Melanoma: A Meta-analysis.}
\format{
A matrix with 70 rows and 15 columns. Each row represents a study's results, the columns are:
\describe{
\item{"TP"}{number of true positives.}
\item{"TN"}{number of ture negatives.}
\item{"FP"}{number of false positives.}
\item{"FN"}{number of false negative.}
\item{"study_ID"}{Study identification}
\item{"test_set_source"}{Public or propietary.}
\item{"method"}{Diagnostic technique used in the study: computer vision; deep learning or hardware-based.}
\item{"test_set_independent"}{yes or no.}
\item{"SAMPLE_SELECTION_BR"}{QUADAS-2, Patient selection bias.}
\item{"INDEX_TEST_BR"}{QUADAS-2, Index test description/application bias.}
\item{"REFERENCE_STANDARD_BR"}{QUADAS-2, Reference standard bias.}
\item{"FLOW_AND_TIMING_BR"}{QUADAS-2, Patient flow and timing bias.}
\item{"SAMPLE_SELECTION_AP}{QUADAS-2, Patient selection bias.}
\item{"INDEX_TEST_AP"}{QUADAS-2, Index test description/application bias.}
\item{"REFERENCE_STANDARD_AP"}{QUADAS-2, Reference standard bias.}
}
}
\source{
The data were obtained from

Dick V, Sinz C, Mittlböck M, Kittler H, Tschandl P. Accuracy of Computer-Aided Diagnosis of Melanoma:
A Meta-analysis. JAMA Dermatol. 2019 Nov 1;155 11:1291-1299.
doi: 10.1001/jamadermatol.2019.1375. PMID: 31215969; PMCID: PMC6584889.
}
\description{
This data frame contains results 70 studies investigated computer-aided diagnosis of melanoma
}
\keyword{datasets}
