% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{skin2}
\alias{skin2}
\title{Accuracy of Computer-Aided Diagnosis of Melanoma: A Comparative Meta-analysis.}
\format{
A matrix with 14 rows and 12 columns. Each row represents a study's results, the columns are:
\describe{
\item{"study"}{Study name and year}
\item{"TP_CAD"}{number of true positives CAD.}
\item{"TN_CAD"}{number of ture negatives CAD.}
\item{"FP_CAD"}{number of false positives CAD.}
\item{"FN_CAD"}{number of false negative CAD.}
\item{"TP_Dermatologists"}{number of true positives Dermatologists.}
\item{"TN_Dermatologists"}{number of ture negatives Dermatologists.}
\item{"FP_Dermatologists"}{number of false positives Dermatologists.}
\item{"FN_Dermatologists"}{number of false negative Dermatologists.}
\item{"test_set_source"}{The database was public domain or propietary.}
\item{"cad_method"}{Computer-Aided Diagnostic Technique: computer vision; deep learning or hardware-based.}
\item{"test_set_independent"}{yes or no.}
}
}
\source{
The data were obtained from

Dick V, Sinz C, Mittlböck M, Kittler H, Tschandl P. Accuracy of Computer-Aided Diagnosis of Melanoma:
A Meta-analysis. JAMA Dermatol. 2019 Nov 1;155 11:1291-1299.
doi: 10.1001/jamadermatol.2019.1375. PMID: 31215969; PMCID: PMC6584889.
}
\description{
This data frame contains results 14 comparative diagnostic studies: CAD vs Dermatologists
}
\keyword{datasets}
