\name{bamdit-package}
\alias{bamdit-package}
\alias{bamdit}
\docType{package}
\title{Bayesian Meta-Analysis of Diagnostic Data}

\description{
Bayesian meta-analysis of diagnostic test data based on a scale mixtures bivariate random-effects model. Summary statistics are based on pooled and predictive sensitivities and specificities. Model checking is used for diagnostic of random-effects' misfit and predictive data is used to assess models departures from observed data.
}

\details{
\tabular{ll}{
Package: \tab bamdit\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-05 \cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}

\author{
Dr. Pablo Emilio Verde \email{pabloemilio.verde@uni-duesseldorf.de}

Arnold Sykosch \email{arnold.sykosch@uni-duesseldorf.de}
}

\references{
Verde P. E. (2010). Meta-analysis of diagnostic test data: A bivariate
Bayesian modeling approach. Statistics in Medicine. 29, 3088-3102.
}

\keyword{
Bayesian modeling
JAGS
Meta-Analysis
Bivariate random-effects
Scale mixtures}

%\examples{
%data(mri)
%mri
%meta <- metadiag(mri)
%summary(meta)
%bamdit.plot(mri,meta)
%}
