% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{ct}
\alias{ct}
\title{Diagnosis of appendicities with computer tomography scans}
\format{
A matrix with 51 rows and 17 columns. Each row represents study results, the columns are:
   \describe{
    \item{tp}{number of true positives.}
    \item{n1}{number of patients with disease.}
    \item{fp}{number of false positives.}
    \item{n2}{number of patients without disease.}
    \item{Author}{First author and year.}
    \item{country}{Country: EU = 1, others/USA = 2.}
    \item{hosp}{Type of hospital: 1 = university, 2 = others.}
    \item{inclus}{Inclusion criteria: 1 = Suspected, 2 = appendectomy.}
    \item{indfind}{Other CT findings included: 1 = no, 2 = yes.}
    \item{design}{Study design: 1 = prospective, 2 = retrospective.}
    \item{contr}{Contrast medium: 1 = no, 2 = yes.}
    \item{localis}{Localisation: 1 = one area, 2 = more than one area.}
    \item{child}{Children included: 1 = no, 2 = yes.}
    \item{fup.na}{Followup: 0 = no, 1 = yes.}
    \item{refer.na}{Valid reference: 0 = no, 1 = yes.}
    \item{sample.na}{Sample: 0 = selected, 1= consecutive/random.}
    \item{gender.na}{Gender, female: 0 = less than 50\%; 1 = more than 50\%.}
    }
}
\source{
The data were obtainded from

Ohmann C, Verde PE, Gilbers T, Franke C, Fuerst G, Sauerland S,
Boehner H. (2006) Systematic review of CT investigation in suspected acute
appendicitis. \emph{Final Report; Coordination Centre for Clinical Trials,
Heinrich-Heine University}. Moorenstr. 5, D-40225 Duesseldorf Germany.
}
\description{
This data frame corresponds to 51 clinical studies reporting the accuracy of computer tomography (CT) scans for the diagnosis of appendicities.
}
\details{
This data frame corresponds to 51 clinical studies reporting
the accuracy of
computer tomography (CT) scans for the diagnosis of appendicities.
}
\references{
Verde P. E. (2010). Meta-analysis of diagnostic test data: A
bivariate Bayesian modeling approach. \emph{Statistics in Medicine}.
\bold{29}, 3088-3102.
}
\keyword{datasets}
