% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{gould}
\alias{gould}
\title{Accuracy of Positron Emission Tomography for Diagnosis of Pulmonary
Nodules and Mass Lesions}
\format{A matrix with 31 rows and 6 columns. Each row represents study
results, the columns are:

\describe{
\item{tp}{number of true positives.}
\item{n1}{number of patients with disease.}
\item{fp}{number of false positives.}
\item{n2}{number of patients without disease.}
\item{author}{first author of the study.}
\item{year}{publication date.}
}}
\source{
The data were obtainded from

Gould MK, Maclean CC, Kuschner WG, Rydzak CE, Owens Dk. (2001) Accuracy of
positron emission tomography for diagnosis of pulmonary nodules and mass
lesions: a meta-analysis. \emph{The Journal of the American Medical
Association};\bold{285}:914-24.
}
\description{
Data from a Meta-Analysis of Studies Quality of FDG-PET for Diagnosis of SPNs and Mass Lesions
}
\keyword{datasets}
