% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{mri}
\alias{mri}
\title{Diagnosis of lymph node metastasis with magnetic resonance imaging}
\format{
A matrix with 10 rows and 4 columns. Each row represents study results, the columns are:
\describe{
\item{tp}{true positives}
\item{n1}{number of patients with disease}
\item{fp}{false positives}
\item{n2}{number of patients without disease}
}
}
\source{
The data were obtained from

 Scheidler J, Hricak H, Yu KK, Subak L, Segal MR. (1997)
 Radiological evaluation of lymph node metastases in
 patients with cervical cancer: a meta-analysis. \emph{The
   Journal of the American Medical Association};
 \bold{278}:1096-1101.
}
\description{
Diagnosis of lymph node metastasis with magnetic resonance imaging
}
\references{
Verde P. E. (2010). Meta-analysis of diagnostic test
 data: A bivariate Bayesian modeling approach.
 \emph{Statistics in Medicine}. \bold{29}, 3088-3102.
}
\keyword{datasets}
