% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotw.R
\name{plotw}
\alias{plotw}
\title{plotw}
\usage{
plotw(m, group = NULL, group.colors = c("blue", "red"))
}
\arguments{
\item{m}{The modelfit.}

\item{group}{Optional argument which has to be a factor of the same length.
as the number of studies in the data. If set, then the plot is colored by groups.}

\item{group.colors}{A character vector with two color names.}
}
\description{
Conflict of evidence plot: this plot displays the posterior distrubution of the study's weights w1 and w1.
These weights indicates potential conflict of evidence of the studies. The weight w1 indicates deviations
with respect to the specificity and w2 to the sensitivity.
}
\details{
The model object must be fitted with the options: re = "sm" and split.w = TRUE
}
\examples{
## execute analysis
\dontrun{
data(mri)
m.mri <- metadiag(mri)
plotw(m=m.mri)
}
}
\seealso{
\code{\link{metadiag}}.
}
\keyword{file}

