% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdit-package.R
\docType{data}
\name{safdar05}
\alias{safdar05}
\title{Diagnosis of Intravascular Device-Related Bloodstream Infection}
\format{
A matrix with 78 rows and 8 columns. Each row represents study
results, the columns are:
\describe{
\item{tp}{number of true positives.}
\item{n1}{number of patients with disease.}
\item{fp}{number of false positives.}
\item{n2}{number of patients without disease.}
\item{author}{first author of the study.}
\item{year}{publication date.}
\item{technique}{diagnostic technique used in the study.}
\item{duration}{duration of catheterization: short term or long term or both.}
}
}
\source{
The data were obtained from

Safdar N, Fine JP, Maki DG. (2005) Meta-analysis: methods for diagnosing
intravascular device-related bloodstream infection.
 \emph{Ann Intern Med.}; \bold{142}:451-66.
}
\description{
Outcome of individual studies evaluating intravascular device-related
bloodstream infection
}
\keyword{datasets}
