% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.metadiag.R
\name{summary.metadiag}
\alias{summary.metadiag}
\title{Generic summary function for metadiag object in bamdit}
\usage{
\method{summary}{metadiag}(object, digits = 3, intervals = c(0.025, 0.5,
  0.975), ...)
}
\arguments{
\item{object}{The object generated by the metadiag function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{intervals}{A numeric vector of probabilities with values in [0,1]. The default value is
intervals = c(0.025, 0.5, 0.975).}

\item{...}{\dots}
}
\description{
Generic summary function for metadiag object in bamdit
}

